requirements.txt
chromadb>=0.4.22
numpy>=1.26.0
fastapi>=0.110.0
httpx>=0.27.0
requests>=2.31.0
openai>=1.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
sentence-transformers>=2.2.2
torch>=2.0.0
uvicorn[standard]>=0.27.0
pytest>=8.0.0
playwright>=1.40.0
python-docx>=1.1.0
